flask-app/requirements.txt
flask
python-dotenv
flask-mysqldb
mysql-connector-python
celery
redis